+++ app/tests/test_controllers.py
@@ -1,12 +1,19 @@
-"""
+"""API tests.
+
 Application unit tests for controllers are defined in this file.
 """
-from webtest import TestApp as WebTestApp, \
-    TestResponse as WebTestResponse
+from unittest import mock
 
-import utils
+import app.utils
+import base64
+import settings
+
 from application import application
 from snapshottest import TestCase
+from webtest import (
+    TestApp as WebTestApp,
+    TestResponse as WebTestResponse
+)
 
 
 class BaseTestCase(TestCase):
@@ -32,7 +39,7 @@ class TestStatusController(BaseTestCase):
     def setUp(self) -> None:
         super().setUp()
 
-    def testHealth(self) -> None:
+    def test_health(self) -> None:
         """Tests the health check endpoint.
 
         :return: None
@@ -47,35 +54,48 @@ class TestTranslatorController(BaseTestCase):
     """Tests the translator controller."""
 
     _body = {
-        "timestamp": "2018-11-01T12:01:01Z",
-        "productCode": "product-1",
-        "parameters": {
-            "name": "The name",
+        'timestamp': '2018-11-01T12:01:01Z',
+        'productCode': 'product-1',
+        'parameters': {
+            'name': 'The name',
         }
     }
 
     def setUp(self):
         super().setUp()
 
-    def testSignatureValidation(self):
+    def test_signature_validation(self):
         """Tests the generating and validating of the signature.
 
         :return: None
         :rtype: None
         """
-        signature = utils.generate_signature(self._body)
+        signature = app.utils.generate_signed_data(
+            self._body,
+            settings.PRIVATE_KEY
+        )
 
-        self.assertTrue(utils.validate_signature(signature, self._body))
+        self.assertTrue(app.utils.validate_signed_data(
+            self._body,
+            signature,
+            settings.PUBLIC_KEY
+        ))
 
-    def testFetch(self):
+    @mock.patch('app.utils.rfc3339')
+    def test_fetch(self, mock_rfc3339):
         """Tests the fetch endpoint.
 
         :return: None
         :rtype: None
         """
-        signature = utils.generate_signature(self._body)
+        mock_rfc3339.return_value = '2019-03-01T08:31:26+00:00'
+        signature = app.utils.generate_signed_data(
+            self._body,
+            settings.PRIVATE_KEY
+        )
         headers = {
-            'X-Pot-Signature': signature
+            'X-Pot-Signature': base64.b64encode(signature).decode("utf-8"),
+            'X-Pot-App': 'bar'
         }
         self._response = self._app.post_json('/fetch',
                                              params=self._body,
@@ -85,23 +105,28 @@ class TestTranslatorController(BaseTestCase):
 
         self.assertMatchSnapshot(self._response.json_body)
 
-    def testFailedSignature(self):
+    def test_failed_signature(self):
         """Tests invalid body against signature.
 
         :return: None
         :rtype: None
         """
-        signature = utils.generate_signature(self._body)
+        signature = app.utils.generate_signed_data(
+            self._body,
+            settings.PRIVATE_KEY
+        )
 
         self.assertFalse(
-            utils.validate_signature(
+            app.utils.validate_signed_data(
+                {
+                    'productCode': 'product-1'
+                },
                 signature,
-                '{"timestamp": "2018-11-01T12:01:01Z", '
-                '"productCode": "product-1"}'
+                settings.PUBLIC_KEY
             )
         )
 
-    def testMissingAttribute(self):
+    def test_missing_attribute(self):
         """Tests missing mandatory attribute.
 
         :return: None
@@ -113,10 +138,13 @@ class TestTranslatorController(BaseTestCase):
                 'name': 'Test'
             }
         }
-
-        signature = utils.generate_signature(params)
+        signature = app.utils.generate_signed_data(
+            params,
+            settings.PRIVATE_KEY
+        )
         headers = {
-            'X-Pot-Signature': signature
+            'X-Pot-Signature': base64.b64encode(signature).decode("utf-8"),
+            'X-Pot-App': 'bar'
         }
         self._response = self._app.post_json('/fetch',
                                              params=params,