+++ settings.travis.py
@@ -1,0 +1,10 @@
+from os import environ
+
+# The Platform Of Trust public key. THIS IS ONLY FOR TESTING PURPOSES.
+POT_PUBLIC_KEY = environ.get('POT_PUBLIC_KEY')
+
+# Public key used for verifying signatures. THIS IS ONLY FOR TESTING PURPOSES.
+PUBLIC_KEY = environ.get('PUBLIC_KEY')
+
+# Private key used for signing. THIS IS ONLY FOR TESTING PURPOSES.
+PRIVATE_KEY = environ.get('PRIVATE_KEY')